+++ python/jaeger/fps.py
@@ -13,6 +13,7 @@ import os
 import pathlib
 import warnings
 from dataclasses import dataclass
+from glob import glob
 
 from typing import (
     TYPE_CHECKING,
@@ -29,6 +30,7 @@ from typing import (
 )
 
 import numpy
+from astropy.time import Time
 
 from jaeger import can_log, config, log, start_file_loggers
 from jaeger.can import JaegerCAN
@@ -54,6 +56,8 @@ from jaeger.utils import Poller, PollerList
 
 
 if TYPE_CHECKING:
+    from matplotlib.axes import Axes
+
     from jaeger.target.configuration import BaseConfiguration
 
 
@@ -1098,6 +1102,52 @@ class FPS(BaseFPS["FPS"]):
                 status["ieb"] = await self.ieb.get_status()
 
         return status
+
+    async def save_snapshot(
+        self,
+        path: Optional[str | pathlib.Path] = None,
+        return_axes: bool = False,
+    ) -> str | Axes:
+        """Creates a plot with the current arrangement of the FPS array.
+
+        Parameters
+        ----------
+        path
+            The path where to save the plot. Defaults to
+            ``/data/fps/snapshots/MJD/fps_snapshot_<SEQ>.pdf``.
+        return_axes
+            If `True`, returns the matplotlib axes instead of saving the plot.
+
+        """
+
+        from jaeger.target.tools import get_snapshot
+
+        ax = await get_snapshot()
+
+        if return_axes is True:
+            return ax
+
+        if path is not None:
+            ax.figure.savefig(path)
+            return str(path)
+
+        mjd = int(Time.now().mjd)
+        dirpath = f"/data/fps/snapshots/{mjd}"
+        if not os.path.exists(dirpath):
+            os.makedirs(dirpath)
+
+        path_pattern = dirpath + "/fps_snapshot_*.pdf"
+        files = sorted(glob(path_pattern))
+
+        if len(files) == 0:
+            seq = 1
+        else:
+            seq = int(files[-1].split("_")[-1][0:4]) + 1
+
+        path = path_pattern.replace("*", f"{seq:04d}")
+        ax.figure.savefig(path)
+
+        return path
 
     async def _handle_temperature(self):
         """Handle positioners in low temperature."""